requirements.txt
annotated-types==0.7.0
fhir.resources==8.1.0
fhir_core==1.1.4
pydantic==2.12.5
pydantic_core==2.41.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
kafka-python==2.3.0
